+++ pkg/configs/advisory/event/true_positive.go
@@ -1,0 +1,6 @@
+package event
+
+// TruePositiveDeterminationEvent is an event that indicates that a previously
+// detected vulnerability was acknowledged to be a true positive.
+type TruePositiveDeterminationEvent struct {
+}
